requirements.txt
streamlit
pandas
Pillow
matplotlib
openpyxl